Lesson 10 (#9)
* lesson 10 instructions

* lesson 10 implementation

+++ main/ulp/sensor.h
@@ -10,7 +10,7 @@ typedef enum _STATE{
 } STATE;
 
 typedef struct {
-  double temp_in_f;
+  int temp_in_f;
   STATE state;
 } temp_reading_t;
